+++ UnstuckME/UnstuckMEUserGUI/UnstuckMEUserGUI/SubWindows/AddUserRoleWindow.xaml.cs
@@ -47,17 +47,17 @@ namespace UnstuckMEUserGUI.SubWindows
             if (userID != -1)
             {
                 targetUser = UnstuckME.Server.GetUserInfo(userID, UserEmailTxtBx.Text);
-                if (targetUser.Privileges == (int)Privileges.User)
+                if (targetUser.Privileges == Privileges.User)
                 {
                     isUser = true;
                     UserBtn.IsChecked = true;
                 }
-                else if (targetUser.Privileges == (int)Privileges.Moderator)
+                else if (targetUser.Privileges == Privileges.Moderator)
                 {
                     isModerator = true;
                     ModeratorBtn.IsChecked = true;
                 }
-                else if (targetUser.Privileges == (int)Privileges.Admin)
+                else if (targetUser.Privileges == Privileges.Admin)
                 {
                     isAdmin = true;
                     AdminBtn.IsChecked = true;
